Add server command configuration profiles
This commit replaces the "cmd-line" config option with a "profile"
config option that can be used to define one or more command profiles to
use when executing the Minecraft server.

This commit just sets up the initial functionality for command profiles,
including the reading and parsing of the "profile" lines defined in the
minecontrol.init file.