Improve default profile handling
Improves default profile handling by adding the "default-profile" config
to the init file. This allows the selection of a defined profile as the
default. This is better than having a named profile always be the
default since we can know preserve the embedded profile name regardless
of if the default profile changes. This will preserve server integrity
as newer versions of Minecraft are installed.