Add "profile" property support to client property processing
Adds support for reading a profile from the property string passed to
the "start" protocol message. This allows a client to specify a property
when creating a new server OR when starting an existing one (even though
changing profiles may be dangerous unless you know what you are doing).

Also fixes a bug that has been latent in minecontrold for some time
regarding processing of the protocol message properties for display in
error messages. This problem could have caused segfault/AV problems.